Modify the PATCH to be more explicit.
The PATCH payload should be a JSON object of order item IDs with that have an array containing the quantity to update. If any other value is provided besides quantity, or quantity is missing, an error is thrown. Order items are not saved until the payload has been validated.

An additional route has been added for updating a single order item, versus the blob approach.